fix bug setting reset data

+++ app/src/main/java/com/example/nguyennam/financialbook/settingtab/SettingMain.java
@@ -113,18 +113,20 @@ public class SettingMain extends Fragment implements View.OnClickListener,
     @Override
     public void onFinishResetDialog(boolean isDelete) {
         // reset data
-        BudgetRecyclerViewDAO budgetDAO = new BudgetRecyclerViewDAO(context);
-        budgetDAO.deleteAllBudget();
-        ExpenseDAO expenseDAO = new ExpenseDAO(context);
-        expenseDAO.deleteAllExpense();
-        IncomeDAO incomeDAO = new IncomeDAO(context);
-        incomeDAO.deleteAllIncome();
-        AccountRecyclerViewDAO accountDAO = new AccountRecyclerViewDAO(context);
-        accountDAO.deleteAllAccount();
-        accountDAO.addAccount(new AccountRecyclerView("Ví", "Tiền mặt", "VND", "0", "", "0"));
-        accountDAO.addAccount(new AccountRecyclerView("ATM", "Tài khoản ngân hàng", "VND", "0", "", "0"));
-        accountDAO.addAccount(new AccountRecyclerView("Tiết Kiệm", "Tài khoản tiết kiệm", "VND", "0", "", "0"));
-        FileHelper.clearAllTempFile(context);
-        Toast.makeText(context, R.string.deleteSuccessfully, Toast.LENGTH_SHORT).show();
+        if (isDelete) {
+            BudgetRecyclerViewDAO budgetDAO = new BudgetRecyclerViewDAO(context);
+            budgetDAO.deleteAllBudget();
+            ExpenseDAO expenseDAO = new ExpenseDAO(context);
+            expenseDAO.deleteAllExpense();
+            IncomeDAO incomeDAO = new IncomeDAO(context);
+            incomeDAO.deleteAllIncome();
+            AccountRecyclerViewDAO accountDAO = new AccountRecyclerViewDAO(context);
+            accountDAO.deleteAllAccount();
+            accountDAO.addAccount(new AccountRecyclerView("Ví", "Tiền mặt", "VND", "0", "", "0"));
+            accountDAO.addAccount(new AccountRecyclerView("ATM", "Tài khoản ngân hàng", "VND", "0", "", "0"));
+            accountDAO.addAccount(new AccountRecyclerView("Tiết Kiệm", "Tài khoản tiết kiệm", "VND", "0", "", "0"));
+            FileHelper.clearAllTempFile(context);
+            Toast.makeText(context, R.string.deleteSuccessfully, Toast.LENGTH_SHORT).show();
+        }
     }
 }
